+++ bin/lemonade
@@ -1,0 +1,4 @@
+#!/usr/bin/env php
+<?php
+
+require __DIR__ . '/lemonade.php';
