--- bin/swiss-knife
@@ -1,4 +1,0 @@
-#!/usr/bin/env php
-<?php
-
-require __DIR__ . '/swiss-knife.php';
