+++ composer.json
@@ -1,9 +1,9 @@
 {
-    "name": "rector/swiss-knife",
-    "description": "Swiss knife in pocket of every upgrade architect",
+    "name": "tomasvotruba/lemonade",
+    "description": "When life gives you 100 Symfony configs, make single  one",
     "license": "MIT",
     "bin": [
-        "bin/swiss-knife"
+        "bin/lemonade"
     ],
     "require": {
         "php": ">=8.2",
@@ -25,7 +25,7 @@
     },
     "autoload": {
         "psr-4": {
-            "Rector\\SwissKnife\\": "src"
+            "TomasVotruba\\Lemonade\\": "src"
         },
         "classmap": [
             "stubs"
@@ -33,13 +33,8 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Rector\\SwissKnife\\Tests\\": "tests"
+            "TomasVotruba\\Lemonade\\Tests\\": "tests"
         }
-    },
-    "replace": {
-        "symfony/polyfill-ctype": "*",
-        "symfony/polyfill-intl-normalizer": "*",
-        "symfony/polyfill-mbstring": "*"
     },
     "config": {
         "sort-packages": true,
