--- src/Command/ValidateFileLengthCommand.php
@@ -1,61 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\SwissKnife\Command;
-
-use Rector\SwissKnife\Finder\FilesFinder;
-use Rector\SwissKnife\Resolver\TooLongFilesResolver;
-use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\InputArgument;
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Style\SymfonyStyle;
-
-final class ValidateFileLengthCommand extends Command
-{
-    public function __construct(
-        private readonly FilesFinder $projectFilesFinder,
-        private readonly SymfonyStyle $symfonyStyle,
-        private readonly TooLongFilesResolver $tooLongFilesResolver
-    ) {
-        parent::__construct();
-    }
-
-    protected function configure(): void
-    {
-        $this->setName('validate-file-length');
-
-        $this->setDescription('Make sure the file path length are not breaking normal Windows max length');
-        $this->addArgument('sources', InputArgument::REQUIRED | InputArgument::IS_ARRAY, 'Path to project');
-    }
-
-    protected function execute(InputInterface $input, OutputInterface $output): int
-    {
-        /** @var string[] $sources */
-        $sources = (array) $input->getArgument('sources');
-
-        $fileInfos = $this->projectFilesFinder->find($sources);
-        $tooLongFileInfos = $this->tooLongFilesResolver->resolve($fileInfos);
-
-        if ($tooLongFileInfos === []) {
-            $message = sprintf('Checked %d files - all fit max file length', count($fileInfos));
-            $this->symfonyStyle->success($message);
-
-            return self::SUCCESS;
-        }
-
-        foreach ($tooLongFileInfos as $tooLongFileInfo) {
-            $message = sprintf(
-                'Paths for file "%s" has %d chars, but must be shorter than %d.',
-                $tooLongFileInfo->getRealPath(),
-                strlen($tooLongFileInfo->getRealPath()),
-                TooLongFilesResolver::MAX_FILE_LENGTH
-            );
-
-            $this->symfonyStyle->warning($message);
-        }
-
-        return self::FAILURE;
-    }
-}
