+++ src/FileSystem/PathHelper.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\SwissKnife\FileSystem;
+namespace TomasVotruba\Lemonade\FileSystem;
 
 final class PathHelper
 {
