--- tests/Comments/Fixture/some_commented_code.php.inc
@@ -1,5 +1,0 @@
-<?php
-
-// some
-// more
-// lines
