--- tests/EntityClassResolver/Fixture/Entity/SomeEntity.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Rector\SwissKnife\Tests\EntityClassResolver\Fixture\Entity;
-
-class SomeEntity
-{
-}
