--- tests/EntityClassResolver/Fixture/SomeClass.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Rector\SwissKnife\Tests\EntityClassResolver\Fixture;
-
-final class SomeClass
-{
-}
