--- tests/EntityClassResolver/Fixture/config/some.orm.yaml
@@ -1,2 +1,0 @@
-App\Some\Entity\Conference:
-    type: entity
