--- tests/Git/ConflictResolver/Fixture/correct_file.txt
@@ -1,4 +1,0 @@
-a
-b
-c
-d
