--- tests/Git/ConflictResolver/Fixture/some_file.txt
@@ -1,2 +1,0 @@
-<<<<<<<<<
-CONFLICT
