--- tests/Git/ConflictResolver/Fixture/some_other_file.txt
@@ -1,6 +1,0 @@
-<<<<<<< HEAD
-this is some content to mess with
-content to append
-=======
-totally different content to merge later
->>>>>>> new_branch_to_merge_later
