--- tests/SomeApp/templates/FirstName/SecondName/template.html.twig
@@ -1,1 +1,0 @@
-{{ hi }}
