--- tests/SomeBundle/RealClassBundle.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\SwissKnife\Tests\SomeBundle;
-
-/**
- * To determine bundle prefix
- */
-final class RealClassBundle
-{
-}
