--- tests/Testing/UnitTestFilePathsFinder/Fixture/IntegrationTest.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\SwissKnife\Tests\Testing\UnitTestFilePathsFinder\Fixture;
-
-use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
-
-final class IntegrationTest extends KernelTestCase
-{
-}
