--- tests/Testing/UnitTestFilePathsFinder/Fixture/OldSchoolTest.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\SwissKnife\Tests\Testing\UnitTestFilePathsFinder\Fixture;
-
-use PHPUnit_Framework_TestCase;
-
-final class OldSchoolTest extends PHPUnit_Framework_TestCase
-{
-}
