--- tests/Testing/UnitTestFilePathsFinder/Fixture/RandomTest.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\SwissKnife\Tests\Testing\UnitTestFilePathsFinder\Fixture;
-
-use PHPUnit\Framework\TestCase;
-
-final class RandomTest extends TestCase
-{
-}
